+++ packages/core/transpiler/view-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/view-parser
 
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- feat: add willParseTemplate support
+
 ## 1.0.0-alpha.2
 
 ### Patch Changes